Github action for building and pushing repo image
PROBLEM:

We need an automated way to build and push a container image of this
repo's contents for every branch update that's pushed. This will allow
us to attach that image to a running k8s pod and ultimately serve out
Puppet environments containint he code changes.

SOLUTION:

Use a GitHub action. Because this process is replacing R10k, we will
also need to handle non-word characters in branch/environment names.
This is being done using the same Ruby `gsub()` call that R10k uses for
compatibility.

OUTCOME:

Changes to any branch of this repo trigger an image build that gets
pushed to `quay.io`.